requirements.txt
# Neo4j
neo4j>=5.0.0
py2neo>=2021.2.0

# Deep Learning
torch>=2.0.0
torch_geometric>=2.3.0
torchvision>=0.15.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utils
pyyaml>=6.0
tqdm>=4.65.0
